refactor(agent): fold vmux_vibe_setup into vmux_agent
Merge the standalone vmux_vibe_setup crate into vmux_agent as the
vibe_setup module. Makes vmux_agent dual-target: native deps gated under
cfg(not(wasm32)) with bevy_ecs/bevy_reflect declared explicitly (bevy
derives emit subcrate paths when bevy is target-gated), dioxus/vmux_ui
added for wasm. Repoint vmux_desktop and vmux_server consumers; delete
the old crate.
