+++ crates/vmux_agent/src/vibe_setup/event.rs
