--- crates/vmux_vibe_setup/Cargo.toml
@@ -1,27 +1,0 @@
-[package]
-name = "vmux_vibe_setup"
-description = "Vibe CLI setup page (vmux://vibe-setup)"
-version.workspace = true
-edition.workspace = true
-publish = false
-
-[features]
-default = []
-web = []
-
-[lib]
-path = "src/lib.rs"
-
-[dependencies]
-rkyv = { workspace = true }
-serde = { workspace = true }
-
-[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
-bevy = { workspace = true }
-bevy_cef = { workspace = true }
-vmux_core = { path = "../vmux_core" }
-vmux_terminal = { path = "../vmux_terminal" }
-
-[target.'cfg(target_arch = "wasm32")'.dependencies]
-dioxus = { workspace = true }
-vmux_ui = { path = "../vmux_ui", default-features = false }
